Confirmacion eliminar usuario de curso

+++ PortalAcademy/src/Controlador/CtrEditarParticipantes.java
@@ -32,12 +32,18 @@ public class CtrEditarParticipantes implements ActionListener {
 	public void actionPerformed(ActionEvent e) {
 
 		if (e.getActionCommand().equals("ELIMINAR")) {
+			
+			
 			usuario = ventana.getUsuario();
 			if (usuario != null) {
-				curso.eliminarUsuario(usuario);
-				JOptionPane.showMessageDialog(ventana, "Se ha eliminado al usuario del curso correctamente");
-				CtrEditarParticipantes c = new CtrEditarParticipantes(new EditarParticipantes(curso, profesor));
-				Main.setPanel(c.getPanel());
+				
+				if (JOptionPane.showConfirmDialog(ventana, "¿Está seguro de que desea eliminar este usuario?", "Eliminar usuario", JOptionPane.OK_CANCEL_OPTION) == 0) {
+					curso.eliminarUsuario(usuario);
+					JOptionPane.showMessageDialog(ventana, "Se ha eliminado al usuario del curso correctamente");
+					CtrEditarParticipantes c = new CtrEditarParticipantes(new EditarParticipantes(curso, profesor));
+					Main.setPanel(c.getPanel());
+				}
+				
 			}
 			
 		}
